backend/requirements.txt
# =============================================================================
# FILE: easycall/backend/requirements.txt
# =============================================================================
# Python dependencies for the Django backend
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Django Core
# -----------------------------------------------------------------------------
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
# SQLite is built into Python - no additional package needed

# -----------------------------------------------------------------------------
# WebSocket Support (Real-time Logging)
# -----------------------------------------------------------------------------
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# -----------------------------------------------------------------------------
# File Processing
# -----------------------------------------------------------------------------
openpyxl>=3.1,<4.0              # Excel files (.xlsx)
pandas>=2.1,<3.0                # Data manipulation
PyPDF2>=3.0,<4.0                # PDF parsing
python-docx>=1.1,<2.0           # Word documents (.docx)
xlrd>=2.0,<3.0                  # Legacy Excel files (.xls)

# -----------------------------------------------------------------------------
# API Clients
# -----------------------------------------------------------------------------
requests>=2.31,<3.0             # HTTP requests to external APIs
httpx>=0.25,<1.0                # Async HTTP client (optional)

# -----------------------------------------------------------------------------
# Security
# -----------------------------------------------------------------------------
cryptography>=41.0,<43.0        # API key encryption
python-dotenv>=1.0,<2.0         # Environment variable management

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8,<3.0       # Date/time utilities
pydantic>=2.5,<3.0              # Data validation

# -----------------------------------------------------------------------------
# Code Quality (Development)
# -----------------------------------------------------------------------------
flake8>=7.0,<8.0                # Linting
black>=24.0,<25.0               # Code formatting
isort>=5.13,<6.0                # Import sorting

# -----------------------------------------------------------------------------
# Testing (Development)
# -----------------------------------------------------------------------------
pytest>=8.0,<9.0                # Testing framework
pytest-django>=4.7,<5.0         # Django integration for pytest
pytest-cov>=4.1,<5.0            # Coverage reporting
pytest-asyncio>=0.23,<1.0       # Async test support
factory-boy>=3.3,<4.0           # Test fixtures

# -----------------------------------------------------------------------------
# Documentation (Development)
# -----------------------------------------------------------------------------
drf-spectacular>=0.27,<1.0      # OpenAPI schema generation